front: move all constants to a single file, add breakpoints there
