+++ front/src/constants/index.js
@@ -1,4 +1,9 @@
+/* Colors */
 // TODO harmonize error colors here
 export const primaryBlue = '#0065DB'
 export const secondaryBlue = '#262C65'
 export const helpColor = '#791A8B'
+
+/* Sizes */
+export const mobileBreakpoint = '672px'
+export const intermediaryBreakpoint = '960px'
